requirements.txt
google-genai
pinecone~=7.3.0
protobuf~=5.29.5